latest integration of asif's code

+++ strategies/xiQuant_strategies/xiquantFuncs.py
@@ -3,7 +3,7 @@
 import numpy
 
 from pyalgotrade import dataseries
-#from pyalgotrade.talibext import indicator
+#from pyalgotrade.talibext import indicator ### blocked to remove ta-lib references....
 
 # Returns the last values of a dataseries as a numpy.array, or None if not enough values
 # could be retrieved from the dataseries.
@@ -12,7 +12,8 @@ def dsToNumpyArray(ds, count):
 	ret = None
 	try:
 		values = ds[count * -1:]
-		ret = numpy.array([float(value) for value in values])
+		#ret = numpy.array([float(value) for value in values], dtype=np.float)
+		ret = numpy.array([value for value in values], dtype=numpy.float)
 	except IndexError:
 		pass
 	except TypeError: # In case we try to convert None to float.
